refactor: move PartitionsSource into data_types.
The compactor currently uses PartitionsSource in two ways:
* for the preparation of PartitionIds prior to the compactor pipeline.
* for the abstraction which utilize the PartitionIds during the IO pipeline.

This commit is a refactoring to enable us to delineate between these two utilizations of the PartitionsSource trait. The prepartion of PartitionsSource is effectively what will be moved into the scheduler.
