+++ compactor/src/components/combos/tests.rs
@@ -1,13 +1,12 @@
 use std::{sync::Arc, time::Duration};
 
-use data_types::{CompactionLevel, PartitionId};
+use data_types::{CompactionLevel, MockPartitionsSource, PartitionId, PartitionsSource};
 use iox_time::{MockProvider, Time};
 
 use crate::components::{
     combos::{throttle_partition::throttle_partition, unique_partitions::unique_partitions},
     commit::{mock::MockCommit, Commit},
     partition_done_sink::{mock::MockPartitionDoneSink, PartitionDoneSink},
-    partitions_source::{mock::MockPartitionsSource, PartitionsSource},
 };
 
 #[tokio::test]
