+++ compactor/src/components/combos/unique_partitions.rs
@@ -7,12 +7,10 @@ use std::{
 };
 
 use async_trait::async_trait;
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use futures::StreamExt;
 
-use crate::components::{
-    partition_done_sink::PartitionDoneSink, partitions_source::PartitionsSource,
-};
+use crate::components::partition_done_sink::PartitionDoneSink;
 
 /// Ensures that a unique set of partitions is flowing through the critical section of the compactor pipeline.
 ///
@@ -179,12 +177,10 @@ where
 
 #[cfg(test)]
 mod tests {
+    use data_types::MockPartitionsSource;
     use std::collections::HashMap;
 
-    use crate::components::{
-        partition_done_sink::mock::MockPartitionDoneSink,
-        partitions_source::mock::MockPartitionsSource,
-    };
+    use crate::components::partition_done_sink::mock::MockPartitionDoneSink;
 
     use super::*;
