+++ compactor/src/components/partition_stream/endless.rs
@@ -1,9 +1,7 @@
 use std::{collections::VecDeque, fmt::Display, sync::Arc};
 
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use futures::{stream::BoxStream, StreamExt};
-
-use crate::components::partitions_source::PartitionsSource;
 
 use super::PartitionStream;
 
@@ -63,7 +61,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::components::partitions_source::mock::MockPartitionsSource;
+    use data_types::MockPartitionsSource;
 
     use super::*;
