+++ compactor/src/components/partition_stream/once.rs
@@ -1,9 +1,7 @@
 use std::{fmt::Display, sync::Arc};
 
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use futures::{stream::BoxStream, StreamExt};
-
-use crate::components::partitions_source::PartitionsSource;
 
 use super::PartitionStream;
 
@@ -49,7 +47,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::components::partitions_source::mock::MockPartitionsSource;
+    use data_types::MockPartitionsSource;
 
     use super::*;
