+++ compactor/src/components/partitions_source/logging.rs
@@ -1,10 +1,8 @@
 use std::fmt::Display;
 
 use async_trait::async_trait;
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use observability_deps::tracing::{info, warn};
-
-use super::PartitionsSource;
 
 #[derive(Debug)]
 pub struct LoggingPartitionsSourceWrapper<T>
@@ -49,9 +47,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use data_types::MockPartitionsSource;
     use test_helpers::tracing::TracingCapture;
-
-    use crate::components::partitions_source::mock::MockPartitionsSource;
 
     use super::*;
