+++ compactor/src/components/partitions_source/metrics.rs
@@ -1,10 +1,8 @@
 use std::fmt::Display;
 
 use async_trait::async_trait;
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use metric::{Registry, U64Counter};
-
-use super::PartitionsSource;
 
 const METRIC_NAME_PARTITIONS_FETCH_COUNT: &str = "iox_compactor_partitions_fetch_count";
 const METRIC_NAME_PARTITIONS_COUNT: &str = "iox_compactor_partitions_count";
@@ -69,9 +67,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use data_types::MockPartitionsSource;
     use metric::assert_counter;
-
-    use crate::components::partitions_source::mock::MockPartitionsSource;
 
     use super::*;
