+++ compactor/src/components/partitions_source/not_empty.rs
@@ -1,10 +1,8 @@
 use std::{fmt::Display, sync::Arc, time::Duration};
 
 use async_trait::async_trait;
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use iox_time::TimeProvider;
-
-use super::PartitionsSource;
 
 #[derive(Debug)]
 pub struct NotEmptyPartitionsSourceWrapper<T>
@@ -56,10 +54,9 @@ where
 
 #[cfg(test)]
 mod tests {
-    use iox_time::{MockProvider, Time};
-
-    use crate::components::partitions_source::mock::MockPartitionsSource;
     use compactor_test_utils::AssertFutureExt;
+    use data_types::MockPartitionsSource;
+    use iox_time::{MockProvider, Time};
 
     use super::*;
