+++ compactor/src/components/partitions_source/randomize_order.rs
@@ -1,10 +1,8 @@
 use std::fmt::Display;
 
 use async_trait::async_trait;
-use data_types::PartitionId;
+use data_types::{PartitionId, PartitionsSource};
 use rand::{rngs::StdRng, seq::SliceRandom, SeedableRng};
-
-use super::PartitionsSource;
 
 #[derive(Debug)]
 pub struct RandomizeOrderPartitionsSourcesWrapper<T>
@@ -48,7 +46,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::components::partitions_source::mock::MockPartitionsSource;
+    use data_types::MockPartitionsSource;
 
     use super::*;
