+++ data_types/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 license.workspace = true
 
 [dependencies]
+async-trait = "0.1.68"
 chrono = { version = "0.4", default-features = false }
 croaring = "0.8.1"
 influxdb-line-protocol = { path = "../influxdb_line_protocol" }
@@ -27,3 +28,4 @@ assert_matches = "1"
 paste = "1.0.12"
 proptest = { version = "1.2.0", default-features = false }
 test_helpers = { path = "../test_helpers" }
+tokio = { version = "1.28", features = ["macros", "net", "parking_lot", "rt-multi-thread", "signal", "sync", "time"] }
